scripts/data_migration/verify_migration.py
# verify_migration.py
import psycopg2

def verify_migration():
    print("Verifying migration results...")
    # conn = psycopg2.connect(...)
    # cur = conn.cursor()
    # cur.execute("SELECT COUNT(*) FROM municipalities;")
    # count = cur.fetchone()[0]
    # expected = 1918
    # if count == expected:
    #     print("SUCCESS: Record count matches.")
    # else:
    #     print(f"FAILURE: Expected {expected}, got {count}")

if __name__ == '__main__':
    verify_migration()
